+++ Changelog.md
@@ -30,12 +30,12 @@
 - Add p1as-observability prometheus Helm chart to PCB
 - Update prometheus-job-exporter and prometheus-json-exporter image paths (v3.0.0) as part of migrating to ping-cloud-common repo
 - Update seal-secret-values.py to handle customSecrets entries for Helm charts
-- Enhance prometheus and prometheus-agent integration, system test
+- Remove metrics-server and kube-state-metrics from PCB
+- Add LOGSTASH_CUSTOMER_PIPELINE_ENABLED feature flag to conditionally deploy logstash-elastic STS in customer-hub accounts
 
 _Changes:_
 
 - [X] PDO-5993 Prometheus: Add p1as-Observability to PCB
-- [X] PDO-6001 Enhance Prometheus tests (unit, integration, system)
 - [X] PDO-8391 Implement KMS changes for EBS
 - [X] PDO-8951 FluentBit: Remove K8s Metadata and Labels
 - [X] PDO-10079 Upgrade cert-manager to version v1.19.1
@@ -64,6 +64,8 @@ _Changes:_
 - [X] PDO-11092 Grafana: Update Dashboards for (Fluent Bit → 2x Logstash)
 - [X] PDO-11104 Update p1as-beluga-lib and seal-secret-values to handle dynamic secrets
 - [X] PDO-11307 Prometheus: Migrate job-exporter and json-exporter to ping-cloud-common repo
+- [X] PDO-11376 Remove metrics-server and kube-state-metrics from PCB
+- [X] PDO-11365 Logstash: Update to make logstash sts (customer pipeline) optional in customer-hub
 
 ### 2.2.1